冲突解决要点：
- StdioClient/HttpClient：采用上游 SSE 按 id 匹配与 64KB stderr 上限，保留 fork 的 16MB 缓冲区护栏与 execFile 无 shell 注入面收敛
- UsageHandlers：接入上游 UsageStatsCache + 目录监听 + dispose 清理，保留 fork 的结果缓存 LRU 上限
- DiffHandlers：diffContentId 白名单校验统一到 isValidDiffContentId
- 前端/webview 共享层：合并 clientId 路由与条件 JSON 往返、ja 语言、容量上限等双方增量
- 适配上游新增测试（媒体工具工作区护栏、execFile mock），同步 CHANGELOG

+++ fast-tavern-main/py-fast-tavern/src/fast_tavern/core/types.py
@@ -32,8 +32,9 @@ class RegexScriptData(TypedDict):
     view: list[RegexView]
     runOnEdit: bool
     macroMode: RegexMacroMode
-    minDepth: int | None
-    maxDepth: int | None
+    # number | null：保留浮点（对齐 TS normalizeRegexes）
+    minDepth: float | None
+    maxDepth: float | None
 
 
 RegexScriptsFileInput = Union[
@@ -63,8 +64,9 @@ class WorldBookEntry(TypedDict):
     key: list[str]
     secondaryKey: list[str]
     selectiveLogic: WorldBookEntrySelectiveLogic
-    order: int
-    depth: int
+    # number：保留浮点不 int 截断（对齐 TS normalizeWorldbooks）
+    order: float
+    depth: float
     position: WorldBookEntryPosition
     role: WorldBookEntryRole | None
     caseSensitive: bool | None
